Common - image magick. Media - ftp. Rollback prompt. Utils. disable gedit. Hide.

+++ run/install-env-media.sh
@@ -1,13 +1,3 @@
-read -p "Продолжить выполнение ${BASH_SOURCE[0]}? (y/n): " answer
-
-if ! [[ "$answer" =~ ^[YyДд] ]]; then
-    echo "Вы выбрали НЕТ. Выход..."
-    return
-fi
-
-
-echo "Вы выбрали ДА. Выполняем действие..."
-
 # Media environment
 
 #sudo add-apt-repository universe
@@ -23,16 +13,21 @@ sudo apt-get update
 
 sudo apt-get install -y vsftpd
 
-sudo cp /etc/vsftpd.conf /etc/vsftpd.conf.back-$(date +"%Y-%m-%d-%H-%M-%S")
+if [ -f /etc/vsftpd.conf ]; then
+  sudo cp /etc/vsftpd.conf /etc/vsftpd.conf.back-$(date +"%Y-%m-%d-%H-%M-%S")
+  sudo sed -i  's|#write_enable=YES|write_enable=YES|' /etc/vsftpd.conf
+  if ! grep -q "mdtm_write=YES" "/etc/vsftpd.conf"; then
+    printf 'mdtm_write=YES\n' | sudo tee -a /etc/vsftpd.conf
+  fi
+#  if grep -q "mdtm" /etc/vsftpd.conf 2>/dev/null; then
+#      echo "Найдены строки с mdtm"
+#  else
+#      echo "Строки с mdtm не найдены"
+#  fi
+  grep --color=auto "mdtm" /etc/vsftpd.conf
+  sudo service vsftpd restart
+fi
 
-sudo sed -i  's|#write_enable=YES|write_enable=YES|' /etc/vsftpd.conf
-
-echo 'mdtm_write=YES
-' | sudo tee -a /etc/vsftpd.conf
-
-cat /etc/vsftpd.conf | grep mdtm
-
-sudo service vsftpd restart
 
 sudo service vsftpd status
